+++ src/dve/core_engine/backends/implementations/spark/rules.py
@@ -46,6 +46,7 @@ from dve.core_engine.backends.metadata.rules import (
 from dve.core_engine.constants import ROWID_COLUMN_NAME
 from dve.core_engine.functions import implementations as functions
 from dve.core_engine.message import FeedbackMessage
+from dve.core_engine.templating import template_object
 from dve.core_engine.type_hints import Messages
 
 
@@ -407,11 +408,15 @@ class SparkStepImplementations(BaseStepImplementations[DataFrame]):
 
         for record in matched.toLocalIterator():
             messages.append(
+                # NOTE: only templates using values directly accessible in record - nothing nested
+                # more complex extraction done in reporting module
                 FeedbackMessage(
                     entity=config.reporting.reporting_entity_override or config.entity_name,
                     record=record.asDict(recursive=True),
                     error_location=config.reporting.legacy_location,
-                    error_message=config.reporting.message,
+                    error_message=template_object(
+                        config.reporting.message, record.asDict(recursive=True)
+                    ),
                     failure_type=config.reporting.legacy_error_type,
                     error_type=config.reporting.legacy_error_type,
                     error_code=config.reporting.code,