+++ src/dve/core_engine/backends/readers/csv.py
@@ -15,10 +15,7 @@ from dve.core_engine.backends.exceptions import (
     FieldCountMismatch,
     MissingHeaderError,
 )
-from dve.core_engine.backends.implementations.duckdb.duckdb_helpers import (
-    get_polars_type_from_annotation,
-)
-from dve.core_engine.backends.utilities import stringify_model
+from dve.core_engine.backends.utilities import get_polars_type_from_annotation, stringify_model
 from dve.core_engine.type_hints import EntityName
 from dve.parser.file_handling import get_content_length, open_stream
 from dve.parser.file_handling.implementations.file import file_uri_to_local_path
@@ -222,7 +219,7 @@ class CSVFileReader(BaseFileReader):
             target_location = file_uri_to_local_path(target_location).as_posix()
         if schema:
             polars_schema: dict[str, pl.DataType] = {  # type: ignore
-                fld.name: get_polars_type_from_annotation(fld.type_)
+                fld.name: get_polars_type_from_annotation(fld.annotation)
                 for fld in stringify_model(schema).__fields__.values()
             }