+++ tests/test_core_engine/test_backends/test_implementations/test_spark/test_data_contract.py
@@ -20,7 +20,9 @@ from dve.core_engine.type_hints import URI
 from dve.core_engine.validation import RowValidator
 from tests.test_core_engine.test_backends.fixtures import (
     nested_all_string_parquet,
+    nested_all_string_parquet_w_errors,
     simple_all_string_parquet,
+    nested_parquet_custom_dc_err_details
 )
 
 
@@ -88,7 +90,7 @@ def test_spark_data_contract_read_and_write_basic_parquet(
     )
 
 
-def test_ddb_data_contract_read_nested_parquet(nested_all_string_parquet):
+def test_spark_data_contract_read_nested_parquet(nested_all_string_parquet):
     # can we read in a stringified parquet and run the data contract on it?
     # more complex file - nested, arrays of structs
     parquet_uri, contract_meta, _ = nested_all_string_parquet
@@ -170,3 +172,68 @@ def test_ddb_data_contract_read_nested_parquet(nested_all_string_parquet):
             ),
         ]
     )
+
+def test_spark_data_contract_custom_error_details(nested_all_string_parquet_w_errors,
+                                                  nested_parquet_custom_dc_err_details):
+    parquet_uri, contract_meta, _ = nested_all_string_parquet_w_errors
+    data_contract = SparkDataContract()
+
+    entity = data_contract.read_parquet(path=parquet_uri)
+    assert entity.count() == 2
+    assert entity.schema == StructType(
+        [
+            StructField("id", StringType()),
+            StructField("strfield", StringType()),
+            StructField("datetimefield", StringType()),
+            StructField(
+                "subfield",
+                ArrayType(
+                    StructType(
+                        [
+                            StructField("id", StringType()),
+                            StructField("substrfield", StringType()),
+                            StructField("subarrayfield", ArrayType(StringType())),
+                        ]
+                    )
+                ),
+            ),
+        ]
+    )
+    # check processes entity
+    contract_dict = json.loads(contract_meta).get("contract")
+    entities: Dict[str, DataFrame] = {
+        "nested_model": entity,
+    }
+    
+    with open(nested_parquet_custom_dc_err_details) as err_dets:
+        custom_error_details = json.load(err_dets)
+
+    dc_meta = DataContractMetadata(
+        reader_metadata={
+            "nested_model": {
+                ".xml": ReaderConfig(
+                    **contract_dict.get("datasets", {})
+                    .get("nested_model", {})
+                    .get("reader_config", {})
+                    .get(".xml")
+                )
+            }
+        },
+        validators={
+            "nested_model": RowValidator(contract_dict,
+                                         "nested_model",
+                                         error_info=custom_error_details)
+        },
+        reporting_fields={"nested_model": ["id"]},
+    )
+
+    entities, messages, stage_successful = data_contract.apply_data_contract(entities, dc_meta)
+    assert stage_successful
+    assert len(messages) == 2
+    messages = sorted(messages, key= lambda x: x.error_code)
+    assert messages[0].error_code == "SUBFIELDTESTIDBAD"
+    assert messages[0].error_message == "subfield id is invalid: subfield.id - WRONG"
+    assert messages[1].error_code == "TESTIDBAD"
+    assert messages[1].error_message == "id is invalid: id - WRONG"
+
+